% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_proc2d_alignment.R
\name{compute_proc2d_alignment}
\alias{compute_proc2d_alignment}
\title{Optimal rotation and scaling alignment to a smooth curve}
\usage{
compute_proc2d_alignment(q, coefs.compl, type, knots, h)
}
\arguments{
\item{q}{complex srv curve with parametrization, needs to be vectorised.
The result of a call to \code{get_model_data_complex}}

\item{coefs.compl}{complex coefficients of smooth curve}

\item{type}{spline degree}

\item{knots}{basis knots}
}
\value{
optimal rotation G and scaling b
}
\description{
Finds optimal rotation and scaling alignment for a discrete open srv curve to a smooth curve
}
