% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastes-package.R
\docType{package}
\name{elastes-package}
\alias{elastes}
\alias{elastes-package}
\title{elastes: Elastic Full Procrustes Means for Sparse and Irregular Planar Curves}
\description{
Provides functions for the computation of functional elastic shape means over sets of open planar curves. The package is particularly suitable for settings where these curves are only sparsely and irregularly observed. It uses a novel approach for elastic shape mean estimation, where planar curves are treated as complex functions and a full Procrustes mean is estimated from the corresponding smoothed Hermitian covariance surface. This is combined with the methods for elastic mean estimation proposed in Steyer, Stöcker, Greven (2022) \doi{10.1111/biom.13706}. See Stöcker et. al. (2022) \href{https://arxiv.org/abs/2203.10522}{arXiv:2203.10522} for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mpff.github.io/elastes/}
  \item \url{https://github.com/mpff/elastes}
  \item Report bugs at \url{https://github.com/mpff/elastes/issues}
}

}
\author{
\strong{Maintainer}: Manuel Pfeuffer \email{manuel.pfeuffer@hu-berlin.de}

Authors:
\itemize{
  \item Lisa Steyer \email{lisa.steyer@hu-berlin.de}
  \item Almond Stoecker \email{almond.stoecker@epfl.ch}
}

}
\keyword{internal}
