% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastes.R
\docType{package}
\name{elastes}
\alias{elastes}
\title{elastes: elastic full Procrustes means for sparse and irregular planar curves.}
\description{
The elastes package provides functions to compute elastic full procrustes means
for collections of curves.
}
\details{
Compute a mean for a set of observed curves: \code{\link{compute_elastic_shape_mean}}
}
