% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticproc2d.R
\docType{package}
\name{elasticproc2d}
\alias{elasticproc2d}
\title{elasticproc2d: elastic full Procrustes means for sparse and irregular planar curves.}
\description{
The elasticproc2d package provides functions to compute elastic full procrustes means
for collections of curves.
}
\section{Main functions}{

Compute a mean for a set of observed curves: \code{\link{compute_elastic_proc2d_mean}}
}

