% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment_proc2d.R
\name{fit_alignment_proc2d}
\alias{fit_alignment_proc2d}
\title{Optimal rotation and scaling alignment to a smooth curve}
\usage{
fit_alignment_proc2d(
  q,
  type,
  knots,
  var_type,
  coefs.compl,
  method,
  cov_fit,
  pca,
  L
)
}
\arguments{
\item{q}{complex srv curve with parametrization, needs to be vectorised.
The result of a call to \code{get_model_data_complex}}

\item{type}{spline degree}

\item{knots}{basis knots}

\item{var_type}{either "smooth" or "constant" measurement error in cov_fit object}

\item{coefs.compl}{complex coefficients of smooth curve}

\item{method}{temp}

\item{cov_fit}{temp}

\item{pca}{temp}

\item{L}{temp}
}
\value{
optimal rotation G and scaling b
}
\description{
Finds optimal rotation and scaling alignment for a discrete open srv curve to a smooth curve
}
