% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elastic_proc2d_mean.R
\name{get_center}
\alias{get_center}
\title{Calculates position of the curve center as the coordinate-wise mean.}
\usage{
get_center(curve)
}
\arguments{
\item{curve}{curve data}
}
\description{
Calculates position of the curve center as the coordinate-wise mean.
}
