% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_t.R
\name{get_optimal_t}
\alias{get_optimal_t}
\title{Finds optimal alignment for discrete open curves}
\usage{
get_optimal_t(srv_procrustes_curves, coefs, t_optims, type, knots, eps, i)
}
\arguments{
\item{srv_procrustes_curves}{scaling and rotation aligned srv curves}

\item{coefs}{mean coefficients}

\item{t_optims}{current optimal parametrisation}

\item{type}{"smooth" or "polygon"}

\item{knots}{mean basis knots}

\item{eps}{convergence tolerance}

\item{i}{current iteration}
}
\value{
optimal time points for srv_data_curves, without first value 0 and last value 1
optimal time points have the distance of the observation to the srv_curve as an attribute
}
\description{
Finds optimal aligned time points for srv curve q to  srv curve p using
coordinate wise optimisation.
}
