% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elastic_proc2d_mean.R
\name{get_polygon_length}
\alias{get_polygon_length}
\title{Calculates length of curve by connecting its points with line segments.}
\usage{
get_polygon_length(curve)
}
\arguments{
\item{curve}{curve data}
}
\description{
Calculates length of curve by connecting its points with line segments.
}
