% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_procrustes_fit.R
\name{get_procrustes_fit}
\alias{get_procrustes_fit}
\title{Get Procrustes data curve from mean object.}
\usage{
get_procrustes_fit(data_curve)
}
\arguments{
\item{data_curve}{A \code{data.frame} in an \code{elastic_shape_mean} object.}
}
\value{
Aligned data_curve as a \code{data.frame}.
}
\description{
Compute the Procrustes aligned data curve...
}
