% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_procrustes_fit_from_param.R
\name{get_procrustes_fit_from_param}
\alias{get_procrustes_fit_from_param}
\title{Helper functions for calculating procrustes data curve from rotation, scaling and translation parameters.}
\usage{
get_procrustes_fit_from_param(data_curve, rot, scale, trans)
}
\arguments{
\item{data_curve}{A \code{data.frame} with observed points on a curve.
Each row is one point, each variable one coordinate direction. If there is a variable \code{t},
it is treated as the time parametrisation, not as an additional coordinate.}

\item{rot}{The rotation (in radian).}

\item{scale}{The scaling.}

\item{trans}{The translation.}
}
\description{
Compute the procrustes fit given optimal rotation, scaling and translation.
}
