% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_t_discrete.R
\name{optimise_one_coord_analytic}
\alias{optimise_one_coord_analytic}
\title{Does optimisation in one parameter direction}
\usage{
optimise_one_coord_analytic(t, i, r, p, s, q)
}
\arguments{
\item{t}{current time points, first has to be 0, last has to be 1}

\item{i}{index of t that should be updated}

\item{r}{time points for p, first has to be 0, last has to be 1}

\item{p}{square root velocity vectors, one less than time points in r}

\item{s}{time points for q, first has to be 0, last has to be 1}

\item{q}{square root velocity vectors, one less than time points in s}
}
\value{
optimal time points for q with respect to optimisation only in the i-th
coordinate direction
}
\description{
Does optimisation in one parameter direction
}
