% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_polygon_length}
\alias{get_polygon_length}
\title{Calculate the polygon length of a curve}
\usage{
get_polygon_length(curve)
}
\arguments{
\item{curve}{a \code{data.frame} with observed points in each row. Each
variable is one coordinate direction. If there is a variable \code{t},
\code{t_optim} or \code{id}, it is treated as the time parametrization, not
as an additional coordinate.}
}
\value{
The length of \code{curve}, treating it as a polygon.
}
\description{
Calculate the polygon length of a curve
}
