% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_procrustes_fit_from_param.R
\name{get_procrustes_fit}
\alias{get_procrustes_fit}
\title{Helper functions for calculating procrustes data curve from rotation, scaling and translation parameters.}
\usage{
get_procrustes_fit(data_curve)
}
\arguments{
\item{data_curve}{A \code{data.frame} in an \code{elastic_proc2d_mean} object.}
}
\description{
Compute the procrustes fit.
}
