% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_procrustes_fit_from_param.R
\name{get_Procrustes_fit_from_param}
\alias{get_Procrustes_fit_from_param}
\alias{get_procrustes_fit_from_param}
\title{Helper functions for calculating Procrustes data curve from rotation, scaling and translation parameters.}
\usage{
get_procrustes_fit_from_param(
  data_curve,
  rot,
  scale,
  plength,
  trans,
  norm_factor
)
}
\arguments{
\item{data_curve}{A \code{data.frame} with observed points on a curve.
Each row is one point, each variable one coordinate direction. If there is a variable \code{t},
it is treated as the time parametrization, not as an additional coordinate.}

\item{rot}{The rotation (in radian).}

\item{scale}{The scaling.}

\item{plength}{The polygon length of the original curve.}

\item{trans}{The translation.}

\item{norm_factor}{The normalization factor from the smooth curve estimate.}
}
\description{
Compute the Procrustes fit given optimal rotation, scaling and translation.
}
