% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elastic_proc2d_mean.R
\name{remove_duplicate}
\alias{remove_duplicate}
\title{Removes duplicate points. Code from \code{elasdics:::remove_duplicate}.}
\usage{
remove_duplicate(data_curve, closed)
}
\arguments{
\item{closed}{TRUE if curve is closed (not implemented yet)}

\item{curve}{curve data}
}
\description{
Removes duplicate points. Code from \code{elasdics:::remove_duplicate}.
}
